Firebase Connect Class Restructured (#88)
* Firebase Connect Class Restructured
