+++ app/src/main/java/com/example/qrgo/firebase/BaseFirebaseConnectManager.java
@@ -1,0 +1,13 @@
+package com.example.qrgo.firebase;
+
+import com.google.firebase.firestore.FirebaseFirestore;
+
+public class BaseFirebaseConnectManager {
+    protected final FirebaseFirestore db;
+
+    public BaseFirebaseConnectManager() {
+        this.db = FirebaseFirestore.getInstance();
+    }
+
+
+}
