+++ app/src/main/java/com/example/qrgo/listeners/OnUserDeleteListener.java
@@ -1,0 +1,5 @@
+package com.example.qrgo.listeners;
+
+public interface OnUserDeleteListener {
+    void onUserDelete(boolean success);
+}
