+++ app/src/main/java/com/example/qrgo/listeners/OnUserProfileDeleteListener.java
@@ -1,0 +1,5 @@
+package com.example.qrgo.listeners;
+
+public interface OnUserProfileDeleteListener {
+    void onUserProfileDelete(boolean success);
+}
